+++ public/js/app.js
@@ -72919,11 +72919,7 @@ var render = function() {
                               ],
                               attrs: { content: notification.data.from.name }
                             },
-                            [
-                              _vm._v(
-                                "@" + _vm._s(notification.data.from.username)
-                              )
-                            ]
+                            [_vm._v(_vm._s(notification.data.from.name))]
                           ),
                           _vm._v(
                             ": " +
@@ -93846,7 +93842,7 @@ Vue.component('show-character', __webpack_require__(/*! ./components/student/Sho
 Vue.component('show-students-student', __webpack_require__(/*! ./components/student/ShowStudentsStudent.vue */ "./resources/js/components/student/ShowStudentsStudent.vue")["default"]);
 Vue.component('show-group-view', __webpack_require__(/*! ./components/student/ShowGroupView.vue */ "./resources/js/components/student/ShowGroupView.vue")["default"]);
 Vue.component('show-student-info', function () {
-  return Promise.all(/*! import() */[__webpack_require__.e(14), __webpack_require__.e(13)]).then(__webpack_require__.bind(null, /*! ./components/student/ShowStudentInfo.vue */ "./resources/js/components/student/ShowStudentInfo.vue"));
+  return Promise.all(/*! import() */[__webpack_require__.e(16), __webpack_require__.e(15)]).then(__webpack_require__.bind(null, /*! ./components/student/ShowStudentInfo.vue */ "./resources/js/components/student/ShowStudentInfo.vue"));
 });
 Vue.component('show-student-map', __webpack_require__(/*! ./components/student/ShowStudentMap.vue */ "./resources/js/components/student/ShowStudentMap.vue")["default"]);
 Vue.component('show-student', __webpack_require__(/*! ./components/student/ShowStudent.vue */ "./resources/js/components/student/ShowStudent.vue")["default"]);
@@ -93855,11 +93851,11 @@ Vue.component('show-card', __webpack_require__(/*! ./components/card/ShowCard.vu
 Vue.component('show-cards', __webpack_require__(/*! ./components/card/ShowCards.vue */ "./resources/js/components/card/ShowCards.vue")["default"]);
 Vue.component('create-card', __webpack_require__(/*! ./components/card/CreateCard.vue */ "./resources/js/components/card/CreateCard.vue")["default"]);
 Vue.component('create-behaviour', function () {
-  return __webpack_require__.e(/*! import() */ 12).then(__webpack_require__.bind(null, /*! ./components/behaviour/CreateBehaviour.vue */ "./resources/js/components/behaviour/CreateBehaviour.vue"));
+  return __webpack_require__.e(/*! import() */ 14).then(__webpack_require__.bind(null, /*! ./components/behaviour/CreateBehaviour.vue */ "./resources/js/components/behaviour/CreateBehaviour.vue"));
 });
 Vue.component('show-behaviours', __webpack_require__(/*! ./components/behaviour/ShowBehaviours.vue */ "./resources/js/components/behaviour/ShowBehaviours.vue")["default"]);
 Vue.component('create-badge', function () {
-  return __webpack_require__.e(/*! import() */ 11).then(__webpack_require__.bind(null, /*! ./components/badge/CreateBadge.vue */ "./resources/js/components/badge/CreateBadge.vue"));
+  return __webpack_require__.e(/*! import() */ 13).then(__webpack_require__.bind(null, /*! ./components/badge/CreateBadge.vue */ "./resources/js/components/badge/CreateBadge.vue"));
 });
 Vue.component('show-badges', __webpack_require__(/*! ./components/badge/ShowBadges.vue */ "./resources/js/components/badge/ShowBadges.vue")["default"]);
 Vue.component('create-levels', __webpack_require__(/*! ./components/level/CreateLevels.vue */ "./resources/js/components/level/CreateLevels.vue")["default"]);
@@ -93894,7 +93890,7 @@ Vue.component('grade', __webpack_require__(/*! ./components/evaluation/Grade.vue
 Vue.component('report', __webpack_require__(/*! ./components/evaluation/Report.vue */ "./resources/js/components/evaluation/Report.vue")["default"]);
 Vue.component('rubric-management', __webpack_require__(/*! ./components/evaluation/RubricManagement.vue */ "./resources/js/components/evaluation/RubricManagement.vue")["default"]);
 Vue.component('rubric-create', function () {
-  return Promise.all(/*! import() */[__webpack_require__.e(4), __webpack_require__.e(10)]).then(__webpack_require__.bind(null, /*! ./components/evaluation/RubricCreate.vue */ "./resources/js/components/evaluation/RubricCreate.vue"));
+  return Promise.all(/*! import() */[__webpack_require__.e(4), __webpack_require__.e(12)]).then(__webpack_require__.bind(null, /*! ./components/evaluation/RubricCreate.vue */ "./resources/js/components/evaluation/RubricCreate.vue"));
 }); // Utils
 // For use this component the parent component should have a data attibute called "icon", where the icon will be stored
 // Vue.component('icon-selector', () => import('./components/utils/IconSelector.vue').default);
@@ -93914,7 +93910,7 @@ Vue.component('wheel', __webpack_require__(/*! ./components/utils/Wheel.vue */ "
 Vue.component('volume-meter', __webpack_require__(/*! ./components/utils/VolumeMeter.vue */ "./resources/js/components/utils/VolumeMeter.vue")["default"]);
 Vue.component('massive-actions', __webpack_require__(/*! ./components/utils/MassiveActions.vue */ "./resources/js/components/utils/MassiveActions.vue")["default"]);
 Vue.component('chat', function () {
-  return Promise.all(/*! import() */[__webpack_require__.e(15), __webpack_require__.e(16)]).then(__webpack_require__.bind(null, /*! ./components/utils/Chat.vue */ "./resources/js/components/utils/Chat.vue"))["default"];
+  return Promise.all(/*! import() */[__webpack_require__.e(10), __webpack_require__.e(11)]).then(__webpack_require__.bind(null, /*! ./components/utils/Chat.vue */ "./resources/js/components/utils/Chat.vue"))["default"];
 });
 /**
  * Next, we will create a fresh Vue application instance and attach it to